PSR-4 + code cleanup
