+++ composer.json
@@ -1,14 +1,24 @@
 {
     "name": "react/partial",
     "description": "Partial function application.",
-    "keywords": ["partial", "functional-programming"],
+    "keywords": [
+        "partial",
+        "functional-programming"
+    ],
     "license": "MIT",
     "require": {
         "php": ">=5.3.2"
     },
     "autoload": {
-        "psr-0": { "React\\Partial": "src" },
-        "files": ["src/React/Partial/functions.php"]
+        "psr-4": {
+            "React\\Partial\\": "src"
+        },
+        "files": [
+            "src/functions.php"
+        ]
+    },
+    "require-dev": {
+        "phpunit/phpunit": "~4.5"
     },
     "extra": {
         "branch-alias": {
