+++ examples/partial.php
@@ -1,6 +1,6 @@
 <?php
 
-require __DIR__.'/../vendor/autoload.php';
+require __DIR__ . '/../vendor/autoload.php';
 
 use React\Partial;
 
@@ -10,5 +10,4 @@ $add = function ($a, $b) {
 
 $addOne = Partial\bind($add, 1);
 
-echo sprintf("%s\n", $addOne(5));
-// outputs 6
+print $addOne(5) . "\n"; // 1 + 5 = 6
